feat: Create routes for User model
This commit adds the register, login, and get user data (/me) RESTful routes for the User schema in userModels.js. `registerUser` implements password hashing and basic validation. The `loginUser` method is used to authenticate a user via their email address and password, and tokens are generated with `generateToken`. `getMe` is a protected route that returns JSON data about the user who is currently logged in. Finally, `userRouters.js` exports these routes and applies basic rate limiting to `/register` in an effort to mitigate abuse.