feat: Implement user logout functionality
This commit adds the functionality to log out users from the system  by updating and saving the user's token to an empty string. It includes a new `logoutUser` controller and a new route `POST /logout` in user routers file. The commit also refactors and adds error-handling functionality to the `loginUser` controller.